instyle_repair/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.2
Django==4.1.6
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
pycparser==2.21
PyJWT==2.6.0
pyTelegramBotAPI==4.10.0
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
